backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0
pdf2image==1.16.3
anthropic==0.7.8
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
bcrypt==4.0.1
cryptography==41.0.8